% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerRankingBoruta.R
\name{marker_ranking_boruta}
\alias{marker_ranking_boruta}
\title{Marker Ranking ~ Boruta}
\usage{
marker_ranking_boruta(selected_data, path = NULL, n_cells = 5000,
  max_runs = 100, seed_val = 123, verbose = 0)
}
\arguments{
\item{selected_data}{Selected data from the stim_cell_selector function.}

\item{path}{Path to the folder for the figures.}

\item{max_runs}{Maximum runber of runs for the random forest algorithm.
Default is 100.}

\item{seed_val}{Seed value for Boruta. Default is 123.}

\item{verbose}{0, 1, or 2. Default is 0.}
}
\value{
Figures saved in a given path.
                          And returns a tibble with attributes statistics calculated
                          by Boruta.
}
\description{
Function to run Boruta on the stim - cell population combinations
             that passed the Fisher's exact test to rank the markers
             according to their contribution to the stim cell selection.
}
\examples{
\dontrun{
attribute_stats <- marker_ranking_boruta(selected_data, path = NULL,
                                        max_runs = 1000, seed_val = 123,
                                        verbose = FALSE)
}
}
