% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerRankingLR.R
\name{marker_ranking_lr}
\alias{marker_ranking_lr}
\title{Marker Ranking ~ Logistic Regression}
\usage{
marker_ranking_lr(selected_data, path = NULL, lr_max_it = 50,
  verbose = FALSE)
}
\arguments{
\item{selected_data}{Selected data from the stim_cell_selector function.}

\item{path}{Path to the folder for the figures.}

\item{lr_max_it}{Maximum iteration (default = 50) for logistic regression.}

\item{verbose}{Boolean (TRUE/FALSE). Default is FALSE.}
}
\description{
Logistic regression for each marker to test it's contribution in
                        k-means clustering.
}
\examples{
\dontrun{
lr_stats <- marker_ranking_lr(selected_data, path = NULL, lr_max_it = 50, verbose = FALSE)
}
}
