% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kde.R
\name{plot_kde}
\alias{plot_kde}
\title{Plot Individual Marker Distribution Before and After Selection}
\usage{
plot_kde(selected_data, path, verbose = FALSE)
}
\arguments{
\item{selected_data}{List output of the \code{\link{stim_cell_selector}} function.}

\item{path}{Path to the folder to save figures generated by this function.}

\item{verbose}{Logical. To make function more verbose. Default is FALSE.}
}
\value{
Generates plots in the specified folder and returns 0 upon successfull completion.
}
\description{
Plot Individual Marker Distribution Before and After Selection
}
