% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_marker_ranking_heatmap.R
\name{plot_marker_ranking_heatmap}
\alias{plot_marker_ranking_heatmap}
\title{Marker ranking heatmap}
\usage{
plot_marker_ranking_heatmap(marker_ranking)
}
\arguments{
\item{marker_ranking}{Returned list from the \code{\link{marker_ranking_boruta}} function.}
}
\value{
A ComplexHeatmap object
}
\description{
A consolidated heatmap showing the importance scores of all the state markers (X-axis) from
all the stimulation-cell population combinations that passed the Fisher's exact test (Y-axis).
}
\examples{
\donttest{
mapped_data <- HDStIM(chi11$expr_data, chi11$state_markers,
                      chi11$cluster_col, chi11$stim_label,
                      chi11$unstim_label, seed_val = 123, umap = TRUE,
                      umap_cells = 50, verbose = FALSE)

marker_ranking <- marker_ranking_boruta(mapped_data, path = NULL, n_cells = NULL,
                                        max_runs = 1000, seed_val = 123,
                                        verbose = 0)

pht <- plot_marker_ranking_heatmap(marker_ranking)
}
}
