% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sbp.R
\name{plot_sbp}
\alias{plot_sbp}
\title{Plot Stacked Bar Plots explaining k-means clustering and Fisher's Exact test}
\usage{
plot_sbp(selected_data, path, verbose = FALSE)
}
\arguments{
\item{selected_data}{Returned list from the \code{\link{stim_cell_selector}} or
\code{\link{stim_cell_selector_single_marker}} function.}

\item{path}{Path to the folder to save figures generated by this function.}

\item{verbose}{Logical. To make function more verbose. Default is FALSE.}
}
\value{
This function generates figures in the specified path folder and returns 0.
}
\description{
Plot Stacked Bar Plots explaining k-means clustering and Fisher's Exact test
}
