% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{plot_umap}
\alias{plot_umap}
\title{Plot UMAPs for the Selected Cells}
\usage{
plot_umap(selected_data, path, verbose = FALSE)
}
\arguments{
\item{selected_data}{Returned list from the \code{\link{stim_cell_selector}} function.}

\item{path}{Path to the folder to save figures generated by this function.}

\item{verbose}{Logical. To make function more verbose. Default is FALSE.}
}
\value{
This function generates figures in the specified path folder and returns 0.
}
\description{
Plot UMAPs for the Selected Cells
}
