% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsSelectedMarkers.R
\name{plots_selected_markers}
\alias{plots_selected_markers}
\title{SCS Selected Marker Plots Pipeline}
\usage{
plots_selected_markers(original_data, selected_data_path, figure_path,
  verbose = FALSE)
}
\arguments{
\item{original_data}{Original tibble/data.frame what was used in \code{\link{scs_selected_markers}} for
the \code{\link{stim_cell_selector}} function.}

\item{selected_data_path}{Path to the folder containing the \code{*.rds} files from
\code{\link{scs_selected_markers}} function.}

\item{figure_path}{Path to the folder where the figures should be saved.}

\item{verbose}{Logical. To make function more verbose. Default is FALSE.}
}
\value{
Generates plots in the specified folder and returns 0 upon successfull completion.
}
\description{
Metafunction to plot all the three types of diagnostic
             plots on the data (*.rds files) generated by
             \code{scs_selected_markers} function/pipeline.
}
