% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scsSelectedMarkers.R
\name{scs_selected_markers}
\alias{scs_selected_markers}
\title{SCS Selected Marker Pipeline}
\usage{
scs_selected_markers(
  dat,
  selected_state_markers,
  cluster_col,
  stim_lab,
  unstim_lab,
  path,
  anova_cutoff = 0.05,
  seed_val = NULL,
  umap = FALSE,
  umap_cells = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dat}{A tibble with the single cell data. Cells on rows
and variables/markers on columns.}

\item{selected_state_markers}{Output data from \code{state_marker_selection_lmm} function.}

\item{cluster_col}{Label for the column in the single cell data tibble that has the cluster information per cell.}

\item{stim_lab}{A character vector of stim label(s) for the \code{stim_cell_selector} function.}

\item{unstim_lab}{A character of unstim label(s) for the \code{stim_cell_selector} function.}

\item{path}{Path to the directory where the output for this meta function will be saved.}

\item{anova_cutoff}{FDR cutoff (Default is 0.05) to filter selected markers from \code{state_marker_selection_lmm} function.}

\item{seed_val}{Seed value (integer; default is NULL) for \code{\link{kmeans}} clustering in \code{stim_cell_selector_function}.}

\item{umap}{Boolean (T/F) to carry out UMAP on the selected cells. Default is FALSE to skip UMAP calculation.}

\item{umap_cells}{An integer; for calculating UMAPs take a minimum of \code{umap_cells} per cluster
or the total number of cells if the cluster size is smaller than \code{umap_cells}. Default is NULL.}

\item{verbose}{Logical. To make function more verbose. Default is FALSE.}
}
\value{
A tibble with all the combinations for stim type
        and cluster that did not yield more than two
        state markers for the given anova fdr p-value cutoff.
}
\description{
Meta function to run \code{stim_cell_selector} on
             the markers selected by \code{state_marker_selection_lmm}
             for each eligible stim type and cluster combination.
}
\examples{
\dontrun{
no_markers_comb <- scs_selected_markers(chi11_1k$expr_data, selected_state_markers,
                                        chi11_1k$cluster_col, chi11_1k$stim_label,
                                        chi11_1k$unstim_label, path = NULL,
                                        anova_cutoff = 0.05, seed_val = 123,
                                        umap = FALSE, umap_cells = NULL, verbose = FALSE)
}
}
