% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateMakerSelectionLmm.R
\name{state_marker_selection_lmm}
\alias{state_marker_selection_lmm}
\title{State (signalling) Marker Selection Using Linear Mixed Effects Modelling}
\usage{
state_marker_selection_lmm(dat, state_markers, cluster_col, stim_lab,
  unstim_lab, REML = FALSE, verbose = FALSE)
}
\arguments{
\item{dat}{A tibble with the single cell data. Cells on rows
and variables/markers on columns.}

\item{state_markers}{A character vector with the labels of state
markers from the stimulation panel.}

\item{cluster_col}{Name (character) of the column with cluster/cell population names
in the input tibble.}

\item{stim_lab}{A character vector of stim label(s).}

\item{unstim_lab}{A character of unstim label(s).}

\item{REML}{Logical. To use restricted maximum likely hood estimation
in model fitting.}

\item{verbose}{Logical. To make function more verbose. Default is FALSE.}
}
\value{
A tibble with p-value from ANOVA per stim type and state marker.
}
\description{
Function to select state (signalling) markers that best discriminate
             stimulated cells from unstimulated cells per stimulation condition.
             This method is based on a combination of linear mixed effects modelling
             and ANOVA. The selected list of markers can be used in the multi-marker
             stim cell selection function.
}
\examples{
\dontrun{
selected_state_markers <- state_marker_selection_lmm(chi11_1k$expr_data, chi11_1k$state_markers,
                 chi11_1k$cluster_col, chi11_1k$stim_label, chi11_1k$unstim_label,
                 REML = FALSE, verbose = FALSE)
}
}
