% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_K_Fisher.R
\name{plot_K_Fisher}
\alias{plot_K_Fisher}
\title{Diagnostic plots explaining K-means clustering and Fisher's Exact test}
\usage{
plot_K_Fisher(mapped_data, path, verbose = FALSE)
}
\arguments{
\item{mapped_data}{Returned list from the \code{\link{HDStIM}} function.}

\item{path}{Path to the folder to save figures generated by this function.}

\item{verbose}{Logical. To make function more verbose. Default is FALSE.}
}
\value{
This function generates figures in the specified path folder and returns 0.
}
\description{
Diagnostic plots explaining K-means clustering and Fisher's Exact test
}
