% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_kde}
\alias{plot_kde}
\title{Plot Individual Marker Distribution Before and After Selection}
\usage{
plot_kde(original_data, selected_data, state_markers, cluster_col, path,
  verbose = FALSE)
}
\arguments{
\item{original_data}{Original tibble/data.frame what was used in \code{\link{stim_cell_selector}} function.}

\item{selected_data}{List output of the \code{\link{stim_cell_selector}} function.}

\item{state_markers}{A character vector with the state marker labels that was used in \code{\link{stim_cell_selector}}.}

\item{cluster_col}{Column in the tibble with the cluster IDs in the original data.}

\item{path}{Path to the folder to save figures generated by this function.}

\item{verbose}{Logical. To make function more verbose. Default is FALSE.}
}
\value{
Generates plots in the specified folder and returns 0 upon successfull completion.
}
\description{
Plot Individual Marker Distribution Before and After Selection
}
