% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{plot_umap}
\alias{plot_umap}
\title{Diagnostic UMAP plots showing the partitioning of cells into responding and non-responding groups by HDStIM.}
\usage{
plot_umap(mapped_data, path, verbose = FALSE)
}
\arguments{
\item{mapped_data}{Returned list from the \code{\link{HDStIM}} function.}

\item{path}{Path to the folder to save figures generated by this function.}

\item{verbose}{Logical. To make function more verbose. Default is FALSE.}
}
\value{
This function generates figures in the specified path folder and returns 0.
}
\description{
Diagnostic UMAP plots showing the partitioning of cells into responding and non-responding groups by HDStIM.
}
\examples{
\dontrun{
plot_umap(mapped_data, path, verbose = FALSE)
}
}
