% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimcellselector.R
\name{stim_cell_selector}
\alias{stim_cell_selector}
\title{Stim Cell Selector}
\usage{
stim_cell_selector(dat, type_markers, state_markers, cluster_col, stim_lab,
  unstim_lab, seed_val = F)
}
\arguments{
\item{dat}{A tibble with the single cell data. Cells on rows
and variables/markers on columns.}

\item{state_markers}{A character vector with the labels of state
markers from the stimulation panel.}

\item{cluster_col}{Column in the tibble with the cluster IDs.}

\item{stim_lab}{A character of stim label(s).}

\item{unstim_lab}{A character of unstim label(s).}

\item{seed_val}{Seed value for \code{\link{kmeans}} clustering.}
}
\value{
A tibble with the selected cells.
}
\description{
Function to select cells from the stimulated samples that
             have likely responded to the stimulant.
}
